Edit toolbar: start over
The toolbar will be created with Amtk instead in a following commit.

It's really great to get rid of menu_in_toolbar.vala, because it was
tricky to implement. I remember implementing it first in C for the
<= 0.2 versions, it was the first GObject subclass that I created
in C :-) then I struggled to re-implement it in Vala, and now it's
finnaly gone!! \o/