Edit toolbar: create it with Amtk
Amtk >= 5.1.0 is required (currently git master) because AmtkFactory is
used with *detailed* GAction names. Note that Amtk is not a new
dependency, since Tepl already depended on it.

It's better than before for the toolbar items that contain a dropdown
menu. Now it's a single button that contains two images in an horizontal
GtkGrid: the image related to the menu, plus an arrow going downwards.
Before, the two images were in two separate buttons, with the first
button doing nothing when clicked, only the second button with the arrow
showed the menu.

So yes, this small issue is finally fixed.